express-note 9.17

+++ express-note/routes/api.js
@@ -4,15 +4,27 @@ var Note = require('../model/note');
 
 /* GET users listing. */
 router.get('/notes', function(req, res, next) {
-
-    Note.findAll({raw: true}).then(function(notes) {
+    var uid = req.session.user.id;
+    var query = {raw: true};
+    if (req.session.user) {
+        query.where = {
+            uid: uid
+        }
+    }
+    Note.findAll(query).then(function(notes) {
         console.log('获取notes');
         res.send({status: 0, data: notes});
     });
 });
 router.post('/notes/add', function(req, res, next) {
+    if (!req.session.user) {
+        return res.send({status: 1, errorMsg: '请先登录'})
+    }
+
+    var uid = req.session.user.id;
     var note = req.body.note;
-    Note.create({text: note}).then(function() {
+
+    Note.create({text: note, uid: uid}).then(function() {
         console.log('添加notes');
         res.send({status: 0});
     }).catch(function() {
@@ -20,16 +32,28 @@ router.post('/notes/add', function(req, res, next) {
     });
 });
 router.post('/notes/edit', function(req, res, next) {
-    Note.update({text: req.body.note}, {where: {id: req.body.id}}).then(function() {
+    if (!req.session.user) {
+        return res.send({status: 1, errorMsg: '请先登录'})
+    }
+    var uid = req.session.user.id;
+    Note.update({text: req.body.note}, {where: {id: req.body.id, uid: uid}}).then(function() {
         console.log('编辑notes');
         res.send({status: 0});
-    })
+    }).catch(function() {
+        res.send({status: 1, errorMsg: '数据库出错'})
+    });
 });
 router.post('/notes/delete', function(req, res, next) {
-    Note.destroy({where: {id: req.body.id}}).then(function() {
+    if (!req.session.user) {
+        return res.send({status: 1, errorMsg: '请先登录'})
+    }
+    var uid = req.session.user.id;
+    Note.destroy({where: {id: req.body.id, uid: uid}}).then(function() {
         console.log('删除notes');
         res.send({status: 0});
-    })
+    }).catch(function() {
+        res.send({status: 1, errorMsg: '数据库出错'})
+    });
 });
 
 module.exports = router;
